pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gmail-reader"
version = "0.1.0"
description = "A tool to read and manage Gmail emails"
requires-python = ">=3.10"
authors = [
    {name = "Deepan Sankar", email = "deepan.abi@outlook.com"},
]
dependencies = [
    "google-api-python-client>=2.0.0",
    "google-auth-oauthlib>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=22.0",
    "isort>=5.0",
]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.ruff]
line-length = 80
